backend/requirements.txt
uvicorn>=0.17.6
pydantic[dotenv]>=1.9.0
aiosqlite==0.17.0
fastapi-amis-admin>=0.2.0,<0.3.0
fastapi-user-auth>=0.2.0,<0.3.0
fastapi-scheduler>=0.0.6,<0.1.0
loguru>=0.6.0
python-decouple>=3.6
simplejson==3.17.6
greenlet>=1.1.2